Added new classes and traits in the 'parametric' package
This commit introduces new class and trait definitions in the `parametric` package. The main changes include the addition of the `Option`, `Some`, `None`, `Tree`, `InnerNode`, `Leaf`, `Comparable`, and `Foo` classes and traits. These changes extend the functionality and structure of the project's Scala files. Changes were also made to the `Cell` class and the project workspace configurations were updated.